deploy reworking, rewrite all from scratch probably
