+++ helix-cli/Cargo.toml
@@ -6,8 +6,6 @@ edition = "2021"
 [dependencies]
 clap = { version = "4.5.36", features = ["derive"] }
 helixdb = { path = "../helixdb", features = ["full"] }
-# helix-engine = { path = "../helix-engine" }
-# helix-gateway = { path = "../helix-gateway" }
 tempfile = "3.19.1"
 dirs = "6.0.0"
 serde = { version = "1.0.219", features = ["derive"] }
@@ -18,10 +16,10 @@ libc = "0.2.171"
 indicatif = "0.17"
 tokio = { version = "1.41.1", features = ["full"] }
 open = "5.0.1"
+colored = "3.0.0"
 
 [target.'cfg(windows)'.dependencies]
 windows = { version = "0.61.1", features = ["Win32_System_Threading", "Win32_Foundation"] }
-
 
 [profile.release]
 lto = true
