+++ helix-cli/src/args.rs
@@ -44,9 +44,6 @@ pub enum CommandType {
     /// Start a stopped Helix instance
     Start(StartCommand),
 
-    /// Update Helix CLI to the latest version
-    Update(UpdateCommand),
-
     /// Ingest data into Helix
     Ingest(IngestCommand),
 }
@@ -65,9 +62,6 @@ pub struct DeployCommand {
 
     #[clap(short, long, help = "Port to run the instance on")]
     pub port: Option<u16>,
-
-    #[clap(short, long, help = "Should generate python bindings")]
-    pub gen_py: bool,
 }
 
 #[derive(Debug, Args)]
@@ -143,10 +137,6 @@ pub struct StartCommand {
     #[clap(help = "Instance ID to Start")]
     pub instance_id: String,
 }
-
-#[derive(Debug, Args)]
-#[clap(name = "update", about = "Update Helix CLI to the latest version")]
-pub struct UpdateCommand {}
 
 #[derive(Debug, Args)]
 #[clap(name = "ingest", about = "Ingest data into Helix")]
